+++ src/lib/filters/index.ts
@@ -1,4 +1,1 @@
-export * from './DEPRECATED_drop-shadow-filter-component';
-export * from './blend-component';
-export * from './gaussian-blur-component';
-export * from './offset-component';
+export * from './filter-brightness';