+++ src/lib/index.ts
@@ -3,3 +3,4 @@ export * from './axis';
 export * from './bars';
 export * from './points';
 export * from './legend';
+export * from './filters';